+++ includes/views/import-metabox-edit.php
@@ -13,6 +13,9 @@
 		<div class="form-group">
 			<label class="feedzy-sr-only"><?php echo __( 'RSS Feed sources (comma separated URLs or Feed Categories slug)', 'feedzy-rss-feeds' ); ?></label>
 		</div>
+
+		<?php echo $invalid_source_msg; ?>
+
 		<div class="form-group input-group">
 			<div class="feedzy-button-inside">
 				<input type="text" id="feedzy-import-source" title="<?php _e( 'Make sure you validate the feed by using the validate button on the right', 'feedzy-rss-feeds' ); ?>" name="feedzy_meta_data[source]" placeholder="<?php echo __( 'Source', 'feedzy-rss-feeds' ); ?>" class="form-control" value="<?php echo $source; ?>"/>
@@ -323,32 +326,33 @@
 							<?php echo apply_filters( 'feedzy_render_magic_tags', '', apply_filters( 'feedzy_magic_tags_image', array() ), 'import_post_featured_img' ); ?>
 						</div>
 					</div>
-
-					<br/>
-					<br/>
-
-					<div id="feedzy_image_fallback">
-							<small>
-								<?php
-								 _e( 'What image should be used as the fallback if an item does not have an image?', 'feedzy-rss-feeds' );
-								?>
-							</small>
-
-						<div>
-							<!-- do not add a space after the span closing bracket or before the beginning of span closing tag -->
-							<span id="feedzy_image_fallback_span"><?php if ( ! empty( $fallback_image ) ) { ?>
-									<img src="<?php echo $fallback_image; ?>">
-								<?php } ?></span>
-							<input type="hidden" name="feedzy_meta_data[fallback_img]" id="fallback_img" value="<?php echo $fallback_image; ?>">
-							<button type="button" class="btn btn-small btn-add-fields" id="feedzy-media-upload-add"><?php _e( 'Upload image', 'feedzy-rss-feeds' ); ?></button>
-							<button type="button" class="btn btn-small btn-remove-fields" id="feedzy-media-upload-remove"><?php _e( 'Remove image', 'feedzy-rss-feeds' ); ?></button>
-						</div>
-					</div>
 				</div>
 				
 			</div>
 
-			<div class="feedzy-row">
+			<div class="feedzy-row" id="feedzy_image_fallback">
+				<div class="label_description">
+					<label class="feedzy-sr-only" for="f1-post-content"><?php _e( 'Featured Image Fallback', 'feedzy-rss-feeds' ); ?></label><br/>
+					<small>
+						<?php
+						_e( 'What image should be used as the fallback if an item does not have an image?', 'feedzy-rss-feeds' )
+						?>
+					</small>
+				</div>
+				<div class="feedzy-separator dashicons dashicons-leftright"></div>
+				<div class="form-group input-group form_item">
+					<!-- do not add a space after the span closing bracket or before the beginning of span closing tag -->
+					<span id="feedzy_image_fallback_span"><?php if ( ! empty( $fallback_image ) ) { ?>
+							<img src="<?php echo $fallback_image; ?>">
+						<?php } ?></span>
+					<input type="hidden" name="feedzy_meta_data[fallback_img]" id="fallback_img" value="<?php echo $fallback_image; ?>">
+					<button type="button" class="btn btn-small btn-add-fields" id="feedzy-media-upload-add"><?php _e( 'Upload image', 'feedzy-rss-feeds' ); ?></button>
+					<button type="button" class="btn btn-small btn-remove-fields" id="feedzy-media-upload-remove"><?php _e( 'Remove image', 'feedzy-rss-feeds' ); ?></button>
+				</div>
+			</div>
+
+			<div class="feedzy-row <?php echo apply_filters( 'feedzy_upsell_class', '' ); ?>">
+				<?php echo apply_filters( 'feedzy_upsell_content', '' ); ?>
 				<div class="label_description">
 					<label class="feedzy-sr-only" for="f1-post-content"><?php echo __( 'Post Author', 'feedzy-rss-feeds' ); ?></label><br/>
 					<small>
@@ -358,8 +362,7 @@
 					</small>
 				</div>
 				<div class="feedzy-separator dashicons dashicons-leftright"></div>
-				<div class="form-group input-group form_item <?php echo apply_filters( 'feedzy_upsell_class', '' ); ?>">
-					<?php echo apply_filters( 'feedzy_upsell_content', '' ); ?>
+				<div class="form-group input-group form_item">
 					<div>
 						<input id="feedzy-toggle_author_admin" name="feedzy_meta_data[import_link_author_admin]" class="feedzy-toggle feedzy-toggle-round" type="checkbox" value="yes" <?php echo $import_link_author[0]; ?>>
 						<label for="feedzy-toggle_author_admin" class="feedzy-inline"></label>