+++ includes/views/js/import-metabox-edit.js
@@ -62,6 +62,7 @@
 	}
 
 	function update_status() {
+        var toggle = $( this );
 	    var post_id = $( this ).val();
 	    var status = $( this ).is( ':checked' );
 
@@ -73,8 +74,22 @@
 			'status': status
 		};
 
-		// since 2.8 ajaxurl is always defined in the admin header and points to admin-ajax.php
-		$.post( ajaxurl, data, function() {} );
+        showSpinner(toggle);
+        $.ajax({
+            url: ajaxurl,
+            data: data,
+            method: 'POST',
+            success: function(data){
+                if( ! data.success && status ) {
+                    toggle.parents('tr').find('td.feedzy-source').find('.feedzy-error-critical').remove();
+                    toggle.parents('tr').find('td.feedzy-source').append($(data.data.msg));
+                    toggle.prop( 'checked', false );
+                }
+            },
+            complete: function(){
+                hideSpinner(toggle);
+            }
+        });
 		return true;
 	}
 
@@ -288,6 +303,11 @@
     }
 
     function initSummary() {
+        $('tr.type-feedzy_imports').each(function(i, e){
+            var $lastRunData = $(e).find('script.feedzy-last-run-data').html();
+            $($lastRunData).insertAfter(e);
+        });
+
         // pop-ups for informational text
         $( '.feedzy-dialog' ).dialog({
           modal: true,
@@ -311,8 +331,12 @@
         $('.feedzy-run-now').on('click', function(e){
             e.preventDefault();
             var button = $(this);
-            showSpinner(button);
-            button.parent().find('.feedzy-error').remove();
+            button.val(feedzy.i10n.importing);
+
+            var numberRow = button.parents('tr').find('~ tr.feedzy-import-status-row:first').find('td tr:first');
+            numberRow.find('td').hide();
+            numberRow.find('td:first').addClass('feedzy_run_now_msg').attr('colspan', 5).html(feedzy.i10n.importing).show();
+
             $.ajax({
                 url     : ajaxurl,
                 method  : 'post',
@@ -323,8 +347,10 @@
                     _action      : 'run_now'
                 },
                 success: function(data){
-                    hideSpinner(button);
-                    button.after($('<div class="feedzy-error feedzy-error-critical">' + data.data.msg + '</div>'));
+                    numberRow.find('td:first').html(data.data.msg);
+                },
+                complete: function(){
+                    button.val(feedzy.i10n.run_now);
                 }
             });
         });
@@ -354,6 +380,9 @@
                     _action      : 'purge'
                 },
                 success: function(){
+                    location.reload();
+                },
+                complete: function(){
                     hideSpinner(element);
                 }
             });